requirements.txt
requests
beautifulsoup4
pandas
google-generativeai
playwright
